integrate axion release plugin
- remove unused gradle.properties
- rename service impl to provider
- clean up commented unnecessary cruft
- depend on io.twdps.starter.tracing for zipkin tracing
