+++ buildSrc/src/main/groovy/starter.std.java.application-conventions.gradle
@@ -6,20 +6,17 @@ plugins {
     // Apply the java Plugin to add support for Java.
     id 'java'
     id "idea"
-    //id "io.spring.dependency-management"
     id "org.ajoberstar.grgit"
     // Apply the application plugin to add support for building a CLI application in Java.
     id 'application'
     id "org.springframework.boot" apply true
     id 'starter.java.repo-conventions'
     id 'starter.java.build-conventions'
-    id 'starter.java.coordinate-conventions'
     id 'starter.java.container-conventions'
     id 'starter.java.open-tracing-common-conventions'
     id 'starter.java.style-conventions'
     id 'starter.java.swagger-conventions'
     id 'starter.java.test-conventions'
-    id 'starter.java.publish-jar-conventions'
 
 }
