+++ buildSrc/src/main/groovy/starter.std.java.bom-conventions.gradle
@@ -5,11 +5,8 @@
 plugins {
     // Apply the java Plugin to add support for Java.
     id 'java-platform'
-//    id "idea"
     id 'starter.java.repo-conventions'
-    id 'starter.java.coordinate-conventions'
     id 'starter.java.publish-pom-conventions'
-//    id "org.springframework.boot" apply true
 
 }
