+++ buildSrc/src/main/groovy/starter.std.java.library-conventions.gradle
@@ -7,16 +7,13 @@ plugins {
     id 'java'
     id 'java-library'
     id "idea"
-    //id "io.spring.dependency-management"
     id "org.ajoberstar.grgit"
     id 'starter.java.repo-conventions'
-    id 'starter.java.coordinate-conventions'
     id 'starter.java.open-tracing-common-conventions'
     id 'starter.java.build-conventions'
     id 'starter.java.style-conventions'
     id 'starter.java.swagger-conventions'
     id 'starter.java.test-conventions'
     id 'starter.java.publish-jar-conventions'
-    // Apply the java-library plugin for API and implementation separation.
 }
