+++ controller/build.gradle
@@ -8,13 +8,13 @@ dependencies {
     annotationBom platform("io.twdps.starter:starter-bom:${dps_starter_boot_version}")
 
     implementation project(':api')
-    implementation project(':service:impl')
+    implementation project(':service:provider')
     implementation project(':service:spi')
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.boot:spring-boot-starter-actuator'
 
     implementation "io.opentracing.contrib:opentracing-spring-web-starter"
-//    implementation("org.postgresql:postgresql:${postgresql_version}")
+//    implementation('org.postgresql:postgresql')
 
 }
