+++ controller/src/main/java/io/twdps/starter/controller/account/AccountController.java
@@ -6,8 +6,8 @@ import io.twdps.starter.api.account.responses.AccountResponse;
 import io.twdps.starter.api.account.responses.AddAccountResponse;
 import io.twdps.starter.api.responses.ArrayResponse;
 import io.twdps.starter.controller.account.mapper.AccountRequestMapper;
-import io.twdps.starter.spi.account.AccountManager;
-import io.twdps.starter.spi.account.model.Account;
+import io.twdps.starter.service.spi.account.AccountManager;
+import io.twdps.starter.service.spi.account.model.Account;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
