+++ controller/src/test/java/io/twdps/starter/controller/account/AccountControllerTest.java
@@ -5,8 +5,8 @@ import io.twdps.starter.api.account.responses.AccountResponse;
 import io.twdps.starter.api.account.responses.AddAccountResponse;
 import io.twdps.starter.api.responses.ArrayResponse;
 import io.twdps.starter.controller.account.mapper.AccountRequestMapper;
-import io.twdps.starter.spi.account.AccountManager;
-import io.twdps.starter.spi.account.model.Account;
+import io.twdps.starter.service.spi.account.AccountManager;
+import io.twdps.starter.service.spi.account.model.Account;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
