+++ gradle.properties
@@ -1,41 +1,18 @@
-module_group=io.twdps
-module_artifact=springboot-starter
-module_version=0.1.0-SNAPSHOT
+group=io.twdps.starterkit
 mavenRepository=https://maven.pkg.github.com/ThoughtWorks-DPS/dps-multi-module-starterkit-java
 mavenSnapshotRepository=https://maven.pkg.github.com/ThoughtWorks-DPS/dps-multi-module-starterkit-java
 packaging=jar
 dockerRegistry=twdps
-assertj_core_version=3.11.1
 checkstyle_version=8.36
-cucumber_version=5.3.0
-dps_starter_boot_version=0.0.1-SNAPSHOT
-gradle_intellij_plugin_version=0.4.14
-h2_version=1.4.200
-jackson_version=2.9.9
-jacoco_version=0.8.5
-junit_jupiter_api_version=5.6.2
-lombok_version=1.18.16
-mapstruct_version=1.4.1.Final
-mockito_version=3.6.0
-mockserver_netty_version=5.8.1
-opentracing_mock_version=0.33.0
-opentracing_spring_cloud_jdbc_starter_version=0.3.12
-opentracing_spring_web_starter_version=3.0.1
-postgresql_version=9.4-1206-jdbc42
-spock_core_version=1.3-groovy-2.5
-spring_boot_web_starter_version=3.0.1
+dps_starter_boot_version=0.1.0
 spring_boot_version=2.4.2
-brave_opentracing_version=0.37.3
-opentracing_api_version=0.33.0
-zipkin_reporter_version=2.15.0
-zipkin_sender_okhttp3_version=2.15.0
 
-spring_dependency_management_version=1.0.10.RELEASE
 grgit_version=4.1.0
 swaggerhub_version=1.0.1
 spotless_version=5.7.0
-palantir_docker_version=0.25.0
-palantir_docker_run_version=0.25.0
+//palantir_docker_version=0.25.0
+//palantir_docker_run_version=0.25.0
 talaiot_version=1.4.1
 talaiot_plugin_base_version=1.4.1
 buildtimetracker_version=0.11.0
+axion_release_plugin_version=1.12.1
