--- persistence/model/gradle.properties
@@ -1,1 +1,0 @@
-module_artifact=persistence-model
