--- service/spi/gradle.properties
@@ -1,1 +1,0 @@
-module_artifact=service-spi
