+++ service/spi/src/main/java/io/twdps/starter/service/spi/activator/Activator.java
@@ -1,0 +1,6 @@
+package io.twdps.starter.service.spi.activator;
+
+public interface Activator {
+
+
+}
