--- service/spi/src/main/java/io/twdps/starter/spi/activator/Activator.java
@@ -1,6 +1,0 @@
-package io.twdps.starter.spi.activator;
-
-public interface Activator {
-
-
-}
