+++ settings.gradle
@@ -8,5 +8,10 @@
  */
 
 rootProject.name = 'springboot-starter'
-include('api', 'controller', 'persistence:model','persistence:impl','service:impl','service:spi', 'app')
-//, 'boot:tracing'
+include ':api'
+include ':controller'
+include ':persistence:model'
+include ':persistence:impl'
+include ':service:provider'
+include ':service:spi'
+include ':app'
